Fix MPS inference path
- RoPE: cast cos/sin to q.dtype to avoid mixed-dtype matmul error
- CLI build_model: convert model to bf16 before weight loading (halves peak memory ~13→6.5 GB)
- New snapmind/core/device.py: resolve_device() and get_model_device()

+++ snapmind/layers/positional/rope.py
@@ -49,11 +49,11 @@ class RotaryPositionalEncoding(PositionalEncodingABC):
     ) -> tuple[torch.Tensor, torch.Tensor]:
         if position_ids is None:
             seq_len = q.shape[-2]
-            cos = self.cos_cached[:seq_len].unsqueeze(0).unsqueeze(0)
-            sin = self.sin_cached[:seq_len].unsqueeze(0).unsqueeze(0)
+            cos = self.cos_cached[:seq_len].unsqueeze(0).unsqueeze(0).to(q.dtype)
+            sin = self.sin_cached[:seq_len].unsqueeze(0).unsqueeze(0).to(q.dtype)
         else:
-            cos = self.cos_cached[position_ids].unsqueeze(1)
-            sin = self.sin_cached[position_ids].unsqueeze(1)
+            cos = self.cos_cached[position_ids].unsqueeze(1).to(q.dtype)
+            sin = self.sin_cached[position_ids].unsqueeze(1).to(q.dtype)
         q_rot = q * cos + self._rotate_half(q) * sin
         k_rot = k * cos + self._rotate_half(k) * sin
         return q_rot, k_rot
